refactor: deploy CDA UI and API as separate WARs
