+++ Dockerfile
@@ -1,6 +1,4 @@
 FROM gradle:8.5-jdk11 AS builder
-USER $USER
-RUN --mount=type=cache,target=/home/gradle/.gradle
 WORKDIR /builddir
 COPY . /builddir/
 RUN apt update && apt install -y curl
@@ -12,7 +10,11 @@ SHELL ["/bin/bash", "-c"]
 RUN . "$NVM_DIR/nvm.sh" && nvm install $NODE_VERSION && nvm use $NODE_VERSION
 ENV NODE_PATH=$NVM_DIR/v$NODE_VERSION/lib/node_modules
 ENV PATH=$NVM_DIR/versions/node/v$NODE_VERSION/bin:$PATH
-RUN gradle prepareDockerBuild --info --no-daemon
+# Docker build contexts created from Git worktrees do not include the main
+# repository's Git metadata. Use a deterministic build-only version rather
+# than asking the Gradle Git version plugin to resolve an unavailable worktree.
+RUN --mount=type=cache,target=/home/gradle/.gradle \
+    gradle prepareDockerBuild --info --no-daemon -PversionOverride=docker
 
 FROM alpine:3.23.5 AS tomcat_base
 RUN apk --no-cache upgrade && \
@@ -31,9 +33,6 @@ RUN mkdir /download && \
     mv apache-tomcat-9.0.121 /usr/local/tomcat/ && \
     cd / && \
     rm -rf /usr/local/tomcat/webapps/* && \
-    mkdir /usr/local/tomcat/webapps/ROOT && \
-    printf "<%% response.sendRedirect(\"/cwms-data/\"); %%>\n" > /usr/local/tomcat/webapps/ROOT/index.jsp && \
-        printf "User-agent: *\nAllow: /cwms-data/\nDisallow: /cwms-data/auth/\nDisallow: /cwms-data/catalog/\nDisallow: /cwms-data/timeseries/\nDisallow: /cwms-data/swagger-docs\nDisallow: /auth/\nSitemap: https://cwms-data.usace.army.mil/sitemap.xml\n" > /usr/local/tomcat/webapps/ROOT/robots.txt && \
     mkdir -p /usr/local/tomcat/conf/Catalina/localhost
 # Now replace the Tomcat logging with logback
 # NOTE: I have reviewed this jar in jd-gui and do not see anything malicious, packages are isolated to avoid issues
@@ -52,7 +51,6 @@ CMD ["/usr/local/tomcat/bin/catalina.sh","run"]
 
 FROM tomcat_base AS api
 
-COPY --from=builder /builddir/cda-gui/dist/sitemap.xml /usr/local/tomcat/webapps/ROOT/sitemap.xml
 COPY --from=builder /builddir/cwms-data-api/build/docker/cda/ /usr/local/tomcat
 COPY --from=builder /builddir/cwms-data-api/build/docker/context.xml /usr/local/tomcat/conf
 COPY --from=builder /builddir/cwms-data-api/build/docker/server.xml /usr/local/tomcat/conf
