+++ cda-gui/src/main.jsx
@@ -57,7 +57,7 @@ const router = createBrowserRouter(
       ],
     },
   ],
-  { basename: "/cwms-data" },
+  { basename: "/" },
 );
 
 ReactDOM.createRoot(document.getElementById("root")).render(
