+++ cda-gui/src/main/java/cwms/cda/ui/SpaErrorStatusFilter.java
@@ -1,4 +1,4 @@
-package cwms.cda.servlet;
+package cwms.cda.ui;
 
 import java.io.IOException;
 import java.util.Set;
