+++ cda-gui/src/test/java/cwms/cda/ui/SpaErrorStatusFilterTest.java
@@ -1,4 +1,4 @@
-package cwms.cda.servlet;
+package cwms.cda.ui;
 
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.mockito.Mockito.mock;
@@ -43,7 +43,7 @@ class SpaErrorStatusFilterTest {
         "/user-lists"
     })
     void returnsOkForClientRoutes(String route) throws ServletException, IOException {
-        HttpServletRequest request = buildRequest("GET", "/cwms-data" + route);
+        HttpServletRequest request = buildRequest("GET", route);
         HttpServletResponse response = mock(HttpServletResponse.class);
         FilterChain chain = mock(FilterChain.class);
 
@@ -55,7 +55,7 @@ class SpaErrorStatusFilterTest {
 
     @Test
     void returnsOkForHeadRequest() throws ServletException, IOException {
-        HttpServletRequest request = buildRequest("HEAD", "/cwms-data/swagger-ui");
+        HttpServletRequest request = buildRequest("HEAD", "/swagger-ui");
         HttpServletResponse response = mock(HttpServletResponse.class);
         FilterChain chain = mock(FilterChain.class);
 
@@ -66,8 +66,8 @@ class SpaErrorStatusFilterTest {
     }
 
     @Test
-    void returnsOkForAlternateContextPath() throws ServletException, IOException {
-        HttpServletRequest request = buildRequest("GET", "/spk-data/swagger-ui", "/spk-data");
+    void supportsNonRootContextPath() throws ServletException, IOException {
+        HttpServletRequest request = buildRequest("GET", "/cwms-data-ui/swagger-ui", "/cwms-data-ui");
         HttpServletResponse response = mock(HttpServletResponse.class);
         FilterChain chain = mock(FilterChain.class);
 
@@ -79,7 +79,7 @@ class SpaErrorStatusFilterTest {
 
     @Test
     void preservesNotFoundStatusForUnknownRoutes() throws ServletException, IOException {
-        HttpServletRequest request = buildRequest("GET", "/cwms-data/not-a-client-route");
+        HttpServletRequest request = buildRequest("GET", "/not-a-client-route");
         HttpServletResponse response = mock(HttpServletResponse.class);
         FilterChain chain = mock(FilterChain.class);
 
@@ -91,7 +91,7 @@ class SpaErrorStatusFilterTest {
 
     @Test
     void preservesNotFoundStatusForNonPageRequests() throws ServletException, IOException {
-        HttpServletRequest request = buildRequest("POST", "/cwms-data/swagger-ui");
+        HttpServletRequest request = buildRequest("POST", "/swagger-ui");
         HttpServletResponse response = mock(HttpServletResponse.class);
         FilterChain chain = mock(FilterChain.class);
 
@@ -102,7 +102,7 @@ class SpaErrorStatusFilterTest {
     }
 
     private HttpServletRequest buildRequest(String method, String requestUri) {
-        return buildRequest(method, requestUri, "/cwms-data");
+        return buildRequest(method, requestUri, "");
     }
 
     private HttpServletRequest buildRequest(String method, String requestUri, String contextPath) {
