+++ cwms-data-api/src/main/webapp/WEB-INF/web.xml
@@ -38,10 +38,4 @@
         <filter-name>CorsFilter</filter-name>
         <url-pattern>/*</url-pattern>
     </filter-mapping>
-
-    <!-- Client Side Routing for React -->
-    <error-page>
-        <error-code>404</error-code>
-        <location>/index.html</location>
-    </error-page>
 </web-app>
